api/requirements.txt
asgiref==3.6.0
astroid==2.15.4
backcall==0.2.0
boto3==1.26.123
botocore==1.29.123
certifi==2022.12.7
chardet==3.0.4
decorator==5.1.1
dill==0.3.6
distlib==0.3.6
dj-database-url==2.0.0
Django==3.2
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-extras==0.3
django-filter==23.2
django-heroku==0.3.1
django-storages==1.13.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
environ==1.0
exceptiongroup==1.1.1
filelock==3.10.7
gunicorn==20.1.0
idna==2.10
iniconfig==2.0.0
ipython==7.15.0
isort==5.12.0
jedi==0.18.2
jmespath==1.0.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
ptyprocess==0.7.0
Pygments==2.15.1
PyJWT==2.6.0
pylint==2.17.3
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.23.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.25.11
virtualenv==20.21.0
wcwidth==0.2.6
Werkzeug==2.3.2
whitenoise==6.4.0
wrapt==1.15.0
